services/mock_events/src/requirements.txt
PyYAML==6.0.1
pynubank==2.25.0
kafka-python==2.0.2
